chore: renamed from ansattporten to azurenauth
