+++ frontend/altinn-support-dashboard.client/src/hooks/azureAuthHooks.ts
