+++ frontend/altinn-support-dashboard.client/src/models/azureAuthModels.ts
