+++ frontend/altinn-support-dashboard.client/src/utils/azureAuthApi.ts
